+++ engine/tracer.js
@@ -24,32 +24,32 @@ exports.resourcesUsage = function resourcesUsage() {
 };
 
 exports.resourcesUsage0 = function resourcesUsage0() {
-  if (gIn.params.traceLevel > -1) {
+  if (gT.cLParams.traceLevel > -1) {
     gIn.cLogger.logResourcesUsage('TRC0: ');
   }
 };
 
 // To distinct tracing from common console.log's.
 exports.msg0 = function msg0(msg) {
-  if (gIn.params.traceLevel > -1) {
+  if (gT.cLParams.traceLevel > -1) {
     gIn.cLogger.msgln(`${gIn.tracePrefix}TRC0: ${msg}`);
   }
 };
 
 exports.msg1 = function msg1(msg) {
-  if (gIn.params.traceLevel > 0) {
+  if (gT.cLParams.traceLevel > 0) {
     gIn.cLogger.msgln(`${gIn.tracePrefix}TRC1: ${msg}`);
   }
 };
 
 exports.msg2 = function msg2(msg) {
-  if (gIn.params.traceLevel > 1) {
+  if (gT.cLParams.traceLevel > 1) {
     gIn.cLogger.msgln(`${gIn.tracePrefix}TRC2: ${msg}`);
   }
 };
 
 exports.msg3 = function msg3(msg) {
-  if (gIn.params.traceLevel > 2) {
+  if (gT.cLParams.traceLevel > 2) {
     gIn.cLogger.msgln(`${gIn.tracePrefix}TRC3: ${msg}`);
   }
 };