+++ testsDir/testsSubDir/__tia-tests__/tia/00_sanity/00_checkForGlobals.tia.js
@@ -4,7 +4,6 @@ t.setTitle('Checking that all global variables are exist and correct.');
 
 a.true(typeof gT !== 'undefined', 'gT exists');
 a.true(typeof gIn !== 'undefined', 'gIn exists');
-a.true('config' in gIn, 'gIn.config exists');
+a.true('config' in gT, 'gT.config exists');
 a.true('tInfo' in gIn, 'gIn.tInfo exists');
 a.true('data' in gIn.tInfo, 'gIn.tInfo.data exists');
-