+++ testsDir/testsSubDir/__tia-tests__/tia/09_extLog/00_extLog.tia.js
@@ -6,4 +6,4 @@ let fs = require('fs');
 
 t.setTitle('Test for external log');
 
-fs.writeFileSync(gIn.params.extLog, 'This string must be in test log\n', gT.engineConsts.logEncoding);
+fs.writeFileSync(gT.cLParams.extLog, 'This string must be in test log\n', gT.engineConsts.logEncoding);