+++ utils/mail-utils.js
@@ -58,7 +58,7 @@ const mailOptions = {
  * @returns {Promise<T>}
  */
 exports.send = function send(subj, htmlDif, txtDif, txtAttachments, zipAttachments) {
-  if (!gIn.params.enableEmail) {
+  if (!gT.cLParams.enableEmail) {
     gIn.tracer.msg2('Mail is disabled.');
     return;
   }